build: refactoring project structure
