+++ .idea/gradle.xml
@@ -12,12 +12,15 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/buildSrc" />
             <option value="$PROJECT_DIR$/common" />
             <option value="$PROJECT_DIR$/data" />
             <option value="$PROJECT_DIR$/dataSourceLocal" />
             <option value="$PROJECT_DIR$/dataSourceRemote" />
             <option value="$PROJECT_DIR$/domain" />
             <option value="$PROJECT_DIR$/presenter" />
+            <option value="$PROJECT_DIR$/ui" />
+            <option value="$PROJECT_DIR$/ui/searchmovie" />
           </set>
         </option>
       </GradleProjectSettings>
