+++ settings.gradle
@@ -6,3 +6,4 @@ include ':data'
 include ':dataSourceLocal'
 include ':dataSourceRemote'
 include ':common'
+include ':ui:searchmovie'
