+++ ui/searchmovie/src/androidTest/java/com/example/searchmovie/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.givita.presenter
+package com.example.searchmovie
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.givita.presenter.test", appContext.packageName)
+        assertEquals("com.example.searchmovie.test", appContext.packageName)
     }
 }
